enhance adhock kde language parsing
.split(@).last => m_variant
.split(_).last => m_country

this in particular means that a country defined in the language (such as
en_GB) will for the system locale override the actually country selected.
if you configure klocale to use
  language = en_GB
  country = Australia
you'll still get the system locale en_GB.UTF-8 because the language
outranks the country in this case.
there is but one exception to this rule, and that is en_US.
since en_US essentially is locale=C it can uniformly be applied to
everthing. en_AT essentially would be native english binary strings with
currency and the likes for Austria.
so the way you get en_AU out of klocale=>systemlocale is you configure
klocale to use
  language = en_US
  country = AU
and get a locale of en_AU.

Please note that this is unfortunately workaroundery as I did not find
any better way to check the validity of a locale. Ultimately en_GB +
Australia should still form en_AU, but to make that happen we'd have to
check whether en_AU is a valid system locale. not yet anyway.